+++ eggs/Naaya/trunk/Products/NaayaBase/tests/testNyComments.py
@@ -5,8 +5,10 @@ from Products.Naaya.NyFolder import addNyFolder
 from naaya.content.url.url_item import addNyURL
 from Products.NaayaBase.NyComments import NyComment
 
+
 def physical_path(obj):
     return '/'.join(obj.getPhysicalPath())
+
 
 class NyCommentsTestCase(NaayaFunctionalTestCase):
     def afterSetUp(self):
@@ -46,7 +48,7 @@ class NyCommentsTestCase(NaayaFunctionalTestCase):
     def test_catalog_delete_commented_object(self):
         self.assertEqual(len(self.catalog_search(NyComment.meta_type)), 0)
 
-        comment = self.portal.folder.test_url._comment_add(body='test comment')
+        self.portal.folder.test_url._comment_add(body='test comment')
         self.assertEqual(len(self.catalog_search(NyComment.meta_type)), 1)
 
         self.portal.folder.manage_delObjects(['test_url'])
@@ -56,45 +58,48 @@ class NyCommentsTestCase(NaayaFunctionalTestCase):
         self.login()
         self.assertEqual(len(self.catalog_search(NyComment.meta_type)), 0)
 
-        comment = self.portal.folder.test_url._comment_add(body='test comment')
+        self.portal.folder.test_url._comment_add(body='test comment')
         self.assertEqual(len(self.catalog_search(NyComment.meta_type,
-                             path = physical_path(self.portal.folder))), 1)
+                             path=physical_path(self.portal.folder))), 1)
 
-        addNyFolder(self.portal, 'other_folder', contributor='admin', submission=1)
+        addNyFolder(self.portal, 'other_folder', contributor='admin',
+                    submission=1)
         clipboard = self.portal.folder.manage_copyObjects(['test_url'])
         self.portal.other_folder.manage_pasteObjects(clipboard)
 
         self.assertEqual(len(self.catalog_search(NyComment.meta_type,
-                             path = physical_path(self.portal.other_folder))), 1)
+                             path=physical_path(self.portal.other_folder))), 1)
 
         self.logout()
 
     def test_catalog_cut_commented_object(self):
         self.login()
-        addNyFolder(self.portal, 'other_folder', contributor='admin', submission=1)
+        addNyFolder(self.portal, 'other_folder', contributor='admin',
+                    submission=1)
 
         self.assertEqual(len(self.catalog_search(NyComment.meta_type)), 0)
 
-        comment = self.portal.folder.test_url._comment_add(body='test comment')
+        self.portal.folder.test_url._comment_add(body='test comment')
         self.assertEqual(len(self.catalog_search(NyComment.meta_type,
-                             path = physical_path(self.portal.folder))), 1)
+                             path=physical_path(self.portal.folder))), 1)
 
         clipboard = self.portal.folder.manage_cutObjects(['test_url'])
         self.portal.other_folder.manage_pasteObjects(clipboard)
 
         self.assertEqual(len(self.catalog_search(NyComment.meta_type,
-                             path = physical_path(self.portal.folder))), 0)
+                             path=physical_path(self.portal.folder))), 0)
 
         self.assertEqual(len(self.catalog_search(NyComment.meta_type,
-                             path = physical_path(self.portal.other_folder))), 1)
+                             path=physical_path(self.portal.other_folder))), 1)
 
         self.logout()
 
     def test_get_comments(self):
-        first_comment = self.portal.folder.test_url._comment_add(body='test comment', 
-                                                                 releasedate='2/10/2010')
-        second_comment = self.portal.folder.test_url._comment_add(body='test comment',
-                                                                  releasedate='1/10/2010')
+        self.portal.folder.test_url._comment_add(body='test comment',
+                                                 releasedate='2/10/2010')
+        self.portal.folder.test_url._comment_add(body='test comment',
+                                                 releasedate='1/10/2010')
         self.assertEqual(self.portal.folder.test_url.count_comments(), 2)
-        self.assertEqual(self.portal.folder.test_url.get_comments_list()[0].releasedate,
-                         '1/10/2010')
+        self.assertEqual(
+            self.portal.folder.test_url.get_comments_list()[0].releasedate,
+            '2/10/2010')